+++ lib/eda/api/channel.ex
@@ -46,7 +46,7 @@ defmodule EDA.API.Channel do
   def edit_permissions(channel_id, overwrite_id, opts) do
     body = Map.new(opts)
     # `alow:` would be dropped by Discord, leaving an overwrite that grants nothing.
-    check_options(body, [:allow, :deny, :type], "EDA.API.Channel.edit_permissions/3")
+    check_options!(body, [:allow, :deny, :type], "EDA.API.Channel.edit_permissions/3")
 
     case put("/channels/#{channel_id}/permissions/#{overwrite_id}", body) do
       {:ok, _} -> :ok